# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Industry Evidence Workspace",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/companies": {
      "get": {
        "tags": [
          "companies-v1"
        ],
        "summary": "List Companies",
        "operationId": "list_companies_api_v1_companies_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "instrument-commands-v1"
        ],
        "summary": "Create Company",
        "operationId": "create_company_api_v1_companies_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{company_id}": {
      "get": {
        "tags": [
          "companies-v1"
        ],
        "summary": "Company Dossier",
        "operationId": "company_dossier_api_v1_companies__company_id__get",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          },
          {
            "name": "cutoff",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cutoff"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyDossierResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/companies/{company_id}/stocks": {
      "post": {
        "tags": [
          "instrument-commands-v1"
        ],
        "summary": "Create Stock",
        "operationId": "create_stock_api_v1_companies__company_id__stocks_post",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStockRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/funds": {
      "post": {
        "tags": [
          "instrument-commands-v1"
        ],
        "summary": "Create Fund",
        "operationId": "create_fund_api_v1_funds_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFundRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FundDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/funds/{fund_id}/holding-disclosures": {
      "post": {
        "tags": [
          "instrument-commands-v1"
        ],
        "summary": "Create Holding Disclosure",
        "operationId": "create_holding_disclosure_api_v1_funds__fund_id__holding_disclosures_post",
        "parameters": [
          {
            "name": "fund_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Fund Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHoldingDisclosureRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldingDisclosureDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stocks/{stock_id}/valuation-snapshots": {
      "post": {
        "tags": [
          "instrument-commands-v1"
        ],
        "summary": "Create Valuation Snapshot",
        "operationId": "create_valuation_snapshot_api_v1_stocks__stock_id__valuation_snapshots_post",
        "parameters": [
          {
            "name": "stock_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Stock Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateValuationSnapshotRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValuationSnapshotDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{company_id}/theme-roles": {
      "post": {
        "tags": [
          "instrument-commands-v1"
        ],
        "summary": "Create Theme Role",
        "operationId": "create_theme_role_api_v1_companies__company_id__theme_roles_post",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateThemeRoleRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThemeRoleDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompanyDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "type",
          "created_at"
        ],
        "title": "CompanyDTO"
      },
      "CompanyDossierResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "v1",
            "title": "Schema Version",
            "default": "v1"
          },
          "basis": {
            "$ref": "#/components/schemas/HistoricalBasisDTO"
          },
          "company": {
            "$ref": "#/components/schemas/CompanyIdentityDTO"
          },
          "stocks": {
            "items": {
              "$ref": "#/components/schemas/StockViewDTO"
            },
            "type": "array",
            "title": "Stocks"
          },
          "theme_roles": {
            "items": {
              "$ref": "#/components/schemas/ThemeRoleViewDTO"
            },
            "type": "array",
            "title": "Theme Roles"
          },
          "related_theses": {
            "items": {
              "$ref": "#/components/schemas/RelatedThesisDTO"
            },
            "type": "array",
            "title": "Related Theses"
          },
          "valuations": {
            "items": {
              "$ref": "#/components/schemas/ValuationViewDTO"
            },
            "type": "array",
            "title": "Valuations"
          },
          "fund_holders": {
            "items": {
              "$ref": "#/components/schemas/FundHolderDTO"
            },
            "type": "array",
            "title": "Fund Holders"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "basis",
          "company",
          "stocks",
          "theme_roles",
          "related_theses",
          "valuations",
          "fund_holders"
        ],
        "title": "CompanyDossierResponse"
      },
...
      "CompanyListResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "v1",
            "title": "Schema Version",
            "default": "v1"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CompanyListItemDTO"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "$ref": "#/components/schemas/CursorPage"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "page"
        ],
        "title": "CompanyListResponse"
      },
...
      "CreateCompanyRequest": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "name",
          "type"
        ],
        "title": "CreateCompanyRequest"
      },
      "CreateFundRequest": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "fund_type": {
            "type": "string",
            "title": "Fund Type"
          },
          "scale": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale"
          },
          "establish_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Establish Date"
          },
          "management_company_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Management Company Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "name",
          "fund_type"
        ],
        "title": "CreateFundRequest"
      },
      "CreateHoldingDisclosureRequest": {
        "properties": {
          "stock_id": {
            "type": "string",
            "format": "uuid",
            "title": "Stock Id"
          },
          "weight": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Weight"
          },
          "report_period": {
            "type": "string",
            "format": "date",
            "title": "Report Period"
          },
          "published_at": {
            "type": "string",
            "format": "date-time",
            "title": "Published At"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "stock_id",
          "weight",
          "report_period",
          "published_at",
          "source"
        ],
        "title": "CreateHoldingDisclosureRequest"
      },
      "CreateStockRequest": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "market": {
            "type": "string",
            "title": "Market"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "name",
          "market"
        ],
        "title": "CreateStockRequest"
      },
      "CreateThemeRoleRequest": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "research_case_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Research Case Id"
          },
          "scope": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope"
          },
          "applicable_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applicable From"
          },
          "applicable_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applicable To"
          },
          "source_statement_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Statement Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "role"
        ],
        "title": "CreateThemeRoleRequest"
      },
...
      "CreateValuationSnapshotRequest": {
        "properties": {
          "as_of_date": {
            "type": "string",
            "format": "date",
            "title": "As Of Date"
          },
          "metric_name": {
            "type": "string",
            "title": "Metric Name"
          },
          "metric_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Metric Value"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "definition": {
            "type": "string",
            "title": "Definition"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "as_of_date",
          "metric_name",
          "metric_value",
          "source",
          "definition"
        ],
        "title": "CreateValuationSnapshotRequest"
      },
...
      "FundDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "fund_type": {
            "type": "string",
            "title": "Fund Type"
          },
          "scale": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale"
          },
          "establish_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Establish Date"
          },
          "management_company_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Management Company Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "fund_type",
          "created_at"
        ],
        "title": "FundDTO"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HoldingDisclosureDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "fund_id": {
            "type": "string",
            "format": "uuid",
            "title": "Fund Id"
          },
          "stock_id": {
            "type": "string",
            "format": "uuid",
            "title": "Stock Id"
          },
          "weight": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Weight"
          },
          "report_period": {
            "type": "string",
            "format": "date",
            "title": "Report Period"
          },
          "published_at": {
            "type": "string",
            "format": "date-time",
            "title": "Published At"
          },
          "acquired_at": {
            "type": "string",
            "format": "date-time",
            "title": "Acquired At"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "fund_id",
          "stock_id",
          "weight",
          "report_period",
          "published_at",
          "acquired_at",
          "source",
          "created_at"
        ],
        "title": "HoldingDisclosureDTO"
      },
...
      "StockDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "market": {
            "type": "string",
            "title": "Market"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "company_id",
          "code",
          "name",
          "market",
          "created_at"
        ],
        "title": "StockDTO"
      },
...
      "ThemeRoleDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id"
          },
          "research_case_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Research Case Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "scope": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope"
          },
          "applicable_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applicable From"
          },
          "applicable_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applicable To"
          },
          "source_statement_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Statement Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "company_id",
          "role",
          "created_at"
        ],
        "title": "ThemeRoleDTO"
      },
...
      "ValuationSnapshotDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "stock_id": {
            "type": "string",
            "format": "uuid",
            "title": "Stock Id"
          },
          "as_of_date": {
            "type": "string",
            "format": "date",
            "title": "As Of Date"
          },
          "metric_name": {
            "type": "string",
            "title": "Metric Name"
          },
          "metric_value": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Metric Value"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "definition": {
            "type": "string",
            "title": "Definition"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "stock_id",
          "as_of_date",
          "metric_name",
          "metric_value",
          "source",
          "definition",
          "created_at"
        ],
        "title": "ValuationSnapshotDTO"
      },
...
    }
  }
}